Dummy water requirement predictor

+++ MonitoringSystem/src/datasimulator/sensordatagenerator/impl/LevelSensorDataGenerator.java
@@ -22,7 +22,7 @@ public class LevelSensorDataGenerator extends SensorDataGenerator{
 		int i = 0;
 		
 		for(String property: properties){
-			propertyValueMap.put(property, "" + (low[i] + Math.random() * (high[i] - low[i])));
+			propertyValueMap.put(property, "" + (int)(low[i] +  Math.random() * (high[i] - low[i])));
 			i++;
 		}
